refactor(frontend): split provider model editor
