+++ packages/frontend/src/components/providers/ProviderModelsEditor.tsx
@@ -1,132 +1,9 @@
-import { useState, useEffect } from 'react';
-import {
-  ChevronDown,
-  ChevronRight,
-  Plus,
-  Play,
-  Loader2,
-  CheckCircle,
-  XCircle,
-  Trash2,
-  X,
-  Download,
-  Info,
-  AlertTriangle,
-} from 'lucide-react';
-import { CopyButton } from '../ui/CopyButton';
-import { Button } from '../ui/Button';
-import { Badge } from '../ui/Badge';
-import { OpenRouterSlugInput } from '../ui/OpenRouterSlugInput';
-import { DebouncedInput } from '../ui/DebouncedInput';
-import { Tooltip } from '../ui/Tooltip';
+import { useState } from 'react';
 import type { Provider } from '../../lib/api';
-import { api } from '../../lib/api';
-import { KNOWN_ADAPTERS } from './ProviderAdvancedEditor';
-import { ReasoningRewriteRulesEditor } from './ReasoningRewriteRulesEditor';
-import { apiAccessToKey, hasApiAccess, toggleApiAccess } from '../../lib/apiFormats';
-
-type ApiAccessOption = { type: string; label: string };
-
-const API_ACCESS_OPTIONS: readonly ApiAccessOption[] = [
-  { type: 'chat', label: 'chat' },
-  { type: 'completions', label: 'completions' },
-  { type: 'messages', label: 'messages' },
-  { type: 'gemini', label: 'gemini' },
-  { type: 'responses', label: 'responses' },
-  { type: 'ollama', label: 'ollama' },
-];
-
-// HTTP image protocols — every one needs a reachable base URL, so none of them
-// can be served by an OAuth provider.
-const IMAGE_API_ACCESS_OPTIONS: readonly ApiAccessOption[] = [
-  { type: 'chat', label: 'OpenAI-compatible' },
-  { type: 'openai-images', label: 'OpenAI Images' },
-  { type: 'openrouter-images', label: 'OpenRouter Images' },
-  { type: 'gemini', label: 'Gemini Images' },
-];
-
-// The only image protocol an `openai-codex` OAuth provider can dispatch: the
-// backend rejects any other image target on an OAuth route with a 400.
-const CODEX_IMAGE_API_ACCESS_OPTIONS: readonly ApiAccessOption[] = [
-  { type: 'codex-images', label: 'Codex Images (ChatGPT OAuth)' },
-];
-
-const CODEX_OAUTH_PROVIDER = 'openai-codex';
-const DEFAULT_IMAGE_ACCESS = 'openai-images';
-const CODEX_IMAGE_ACCESS = 'codex-images';
-
-const GPT5_SUPPRESSION_ADAPTER = 'suppress_unsupported_gpt5_options';
-
-function isGpt5Model(modelId: string): boolean {
-  return /^gpt-5(?:[.-]|$)/i.test(modelId);
-}
-
-const getApiBadgeStyle = (apiType: string): React.CSSProperties => {
-  switch (apiType.toLowerCase()) {
-    case 'messages':
-      return { backgroundColor: '#D97757', color: 'white', border: 'none' };
-    case 'chat':
-      return { backgroundColor: '#ebebeb', color: '#333', border: 'none' };
-    case 'completions':
-      return { backgroundColor: '#3b82f6', color: 'white', border: 'none' };
-    case 'gemini':
-      return { backgroundColor: '#5084ff', color: 'white', border: 'none' };
-    case 'embeddings':
-      return { backgroundColor: '#10b981', color: 'white', border: 'none' };
-    case 'transcriptions':
-      return { backgroundColor: '#a855f7', color: 'white', border: 'none' };
-    case 'speech':
-      return { backgroundColor: '#f97316', color: 'white', border: 'none' };
-    case 'openai-images':
-      return { backgroundColor: '#d946ef', color: 'white', border: 'none' };
-    case 'responses':
-      return { backgroundColor: '#06b6d4', color: 'white', border: 'none' };
-    case 'openrouter-images':
-      return { backgroundColor: '#7c3aed', color: 'white', border: 'none' };
-    case 'codex-images':
-      return { backgroundColor: '#10a37f', color: 'white', border: 'none' };
-    case 'ollama':
-      return { backgroundColor: '#1a5f7a', color: 'white', border: 'none' };
-    default:
-      return {};
-  }
-};
-
-// Consistent compact field class used everywhere in the model editor
-const FIELD_CLS =
-  'w-full h-[27px] py-0 px-2 font-body text-[12px] leading-none text-text bg-bg-glass border border-border-glass rounded-sm outline-none focus:border-primary';
-
-// ── ModelIdInputCompact ──
-function ModelIdInputCompact({
-  modelId,
-  onCommit,
-}: {
-  modelId: string;
-  onCommit: (oldId: string, newId: string) => void;
-}) {
-  const [draftId, setDraftId] = useState(modelId);
-  useEffect(() => {
-    setDraftId(modelId);
-  }, [modelId]);
-  const commit = () => {
-    if (!draftId || draftId === modelId) return;
-    onCommit(modelId, draftId);
-  };
-  return (
-    <input
-      className={FIELD_CLS}
-      value={draftId}
-      onChange={(e) => setDraftId(e.target.value)}
-      onBlur={commit}
-      onKeyDown={(e) => {
-        if (e.key === 'Enter') {
-          commit();
-          (e.target as HTMLInputElement).blur();
-        }
-      }}
-    />
-  );
-}
+import { ModelList } from './model-editor/ModelList';
+import { renameRecordKey, removeRecordKey } from './model-editor/adapter-utils';
+import { CODEX_OAUTH_PROVIDER } from './model-editor/constants';
+import { usePiAiModels } from './model-editor/usePiAiModels';
 
 interface Props {
   editingProvider: Provider;
@@ -188,1550 +65,57 @@ export function ProviderModelsEditor({
 }: Props) {
   const [modelAdaptersOpen, setModelAdaptersOpen] = useState<Record<string, boolean>>({});
   const [modelAdvancedOpen, setModelAdvancedOpen] = useState<Record<string, boolean>>({});
+  const piAiModels = usePiAiModels(editingProvider.pi_ai_provider, editingProvider.models);
 
-  // pi-ai model dropdown state — shared across all models (same provider)
-  const [piModels, setPiModels] = useState<
-    Array<{ id: string; name: string; api: string; custom: boolean }>
-  >([]);
-  const [piModelCustom, setPiModelCustom] = useState<Record<string, boolean>>({});
-
-  const piAiProvider = editingProvider.pi_ai_provider;
-
-  // Codex Images rides the provider's ChatGPT OAuth session instead of a base
-  // URL, so it is offered only on an OAuth provider whose backend is Codex —
-  // and there it is the only image protocol the dispatcher will accept.
   const isCodexOAuthProvider =
     isOAuthMode && editingProvider.oauthProvider === CODEX_OAUTH_PROVIDER;
-  const imageAccessOptions = isCodexOAuthProvider
-    ? CODEX_IMAGE_API_ACCESS_OPTIONS
-    : IMAGE_API_ACCESS_OPTIONS;
-  const defaultImageAccess = isCodexOAuthProvider ? CODEX_IMAGE_ACCESS : DEFAULT_IMAGE_ACCESS;
 
-  useEffect(() => {
-    if (!piAiProvider) {
-      setPiModels([]);
-      return;
-    }
-    api
-      .getPiModels(piAiProvider)
-      .then(setPiModels)
-      .catch(() => setPiModels([]));
-  }, [piAiProvider]);
+  const handleUpdateModelId = (oldId: string, newId: string) => {
+    if (oldId === newId) return;
+    setModelAdaptersOpen((prev) => renameRecordKey(prev, oldId, newId));
+    setModelAdvancedOpen((prev) => renameRecordKey(prev, oldId, newId));
+    piAiModels.renameModelId(oldId, newId);
+    updateModelId(oldId, newId);
+  };
 
-  // When models load, mark any existing model IDs that aren't in the registry as custom
-  useEffect(() => {
-    if (piModels.length === 0) return;
-    const modelIds = new Set(piModels.map((m) => m.id));
-    const models = editingProvider.models as Record<string, any> | undefined;
-    if (!models) return;
-    const updates: Record<string, boolean> = {};
-    for (const [mId, mCfg] of Object.entries(models)) {
-      const piId = (mCfg as any).pi_ai_model_id;
-      if (piId && !modelIds.has(piId)) {
-        updates[mId] = true;
-      }
-    }
-    if (Object.keys(updates).length > 0) {
-      setPiModelCustom((prev) => ({ ...prev, ...updates }));
-    }
-  }, [piModels, editingProvider.models]);
+  const handleRemoveModel = (modelId: string) => {
+    setModelAdaptersOpen((prev) => removeRecordKey(prev, modelId));
+    setModelAdvancedOpen((prev) => removeRecordKey(prev, modelId));
+    piAiModels.removeModelId(modelId);
+    removeModel(modelId);
+  };
 
   return (
-    <div className="border border-border-glass rounded-md">
-      <div
-        className="p-2 px-3 flex items-center gap-2 cursor-pointer bg-bg-hover transition-colors duration-200 select-none hover:bg-bg-glass"
-        onClick={() => setIsModelsOpen(!isModelsOpen)}
-      >
-        {isModelsOpen ? <ChevronDown size={16} /> : <ChevronRight size={16} />}
-        <span style={{ fontWeight: 600, fontSize: '13px', flex: 1 }}>Provider Models</span>
-        <Badge status="connected">{Object.keys(editingProvider.models || {}).length} Models</Badge>
-        <Button
-          size="sm"
-          variant="secondary"
-          onClick={(e) => {
-            e.stopPropagation();
-            onOpenFetchModels();
-          }}
-          leftIcon={<Download size={14} />}
-          style={{ marginLeft: '8px' }}
-        >
-          Fetch Models
-        </Button>
-      </div>
-      {isModelsOpen && (
-        <div
-          style={{
-            padding: '8px',
-            borderTop: '1px solid var(--color-border-glass)',
-            background: 'var(--color-bg-subtle)',
-          }}
-        >
-          <div style={{ display: 'flex', flexDirection: 'column', gap: '6px' }}>
-            <Button variant="secondary" size="sm" leftIcon={<Plus size={14} />} onClick={addModel}>
-              Add Model
-            </Button>
-            {Object.entries(editingProvider.models || {}).map(([mId, mCfg]: [string, any]) => {
-              const testKey = `${editingProvider.id}-${mId}`;
-              const testState = testStates[testKey];
-
-              return (
-                <div
-                  key={mId}
-                  style={{
-                    border: '1px solid var(--color-border-glass)',
-                    borderRadius: 'var(--radius-sm)',
-                    background: 'var(--color-bg-surface)',
-                  }}
-                >
-                  <div
-                    style={{
-                      padding: '6px 8px',
-                      display: 'flex',
-                      alignItems: 'center',
-                      gap: '8px',
-                      cursor: 'pointer',
-                    }}
-                    onClick={() => setOpenModelIdx(openModelIdx === mId ? null : mId)}
-                  >
-                    {openModelIdx === mId ? <ChevronDown size={12} /> : <ChevronRight size={12} />}
-                    <span style={{ fontWeight: 600, fontSize: '12px', flex: 1 }}>{mId}</span>
-                    <div
-                      onClick={(e) => {
-                        if (isNewProvider) return;
-                        e.stopPropagation();
-                        onTestModel(editingProvider.id, mId, mCfg.type);
-                      }}
-                      className={
-                        isNewProvider
-                          ? 'flex items-center cursor-not-allowed opacity-40'
-                          : 'flex items-center cursor-pointer'
-                      }
-                      title={
-                        isNewProvider
-                          ? 'Save the provider first to probe models'
-                          : 'Test this model'
-                      }
-                    >
-                      {testState?.loading ? (
-                        <Loader2 size={14} className="animate-spin text-text-secondary" />
-                      ) : testState?.showResult && testState.result === 'success' ? (
-                        <CheckCircle size={14} className="text-success" />
-                      ) : testState?.showResult && testState.result === 'error' ? (
-                        <XCircle size={14} className="text-danger" />
-                      ) : (
-                        <Play size={14} className="text-primary opacity-60" />
-                      )}
-                    </div>
-                    <CopyButton
-                      value={`direct/${editingProvider.id}/${mId}`}
-                      size="sm"
-                      className="mr-1"
-                    />
-                    <Button
-                      size="sm"
-                      variant="ghost"
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        removeModel(mId);
-                      }}
-                      style={{ color: 'var(--color-danger)', padding: '2px' }}
-                    >
-                      <X size={12} />
-                    </Button>
-                  </div>
-                  {testState?.showMessage && testState.result === 'error' && testState.message && (
-                    <div style={{ padding: '0 8px 6px 8px' }}>
-                      <div
-                        onClick={(e) => {
-                          e.stopPropagation();
-                          onDismissTestMessage(testKey);
-                        }}
-                        className="cursor-pointer rounded border border-danger/30 bg-danger/10 px-2 py-1"
-                        title="Click to dismiss"
-                      >
-                        <span className="text-[11px] italic text-danger">
-                          {testState.message} [×]
-                        </span>
-                      </div>
-                    </div>
-                  )}
-
-                  {openModelIdx === mId && (
-                    <div
-                      style={{
-                        padding: '8px',
-                        borderTop: '1px solid var(--color-border-glass)',
-                        display: 'flex',
-                        flexDirection: 'column',
-                        gap: '6px',
-                      }}
-                    >
-                      {/* 2-column primary layout: left = meta, right = pricing */}
-                      <div
-                        style={{
-                          display: 'grid',
-                          gridTemplateColumns: '1fr 1fr',
-                          gap: '8px',
-                          alignItems: 'start',
-                        }}
-                      >
-                        {/* Left: Model ID + Type + Access Via + pi-ai Model ID */}
-                        <div style={{ display: 'flex', flexDirection: 'column', gap: '6px' }}>
-                          {/* Compact Model ID — bypasses Input component's py-2 */}
-                          <div className="flex flex-col gap-1">
-                            <label className="font-body text-[11px] font-medium text-text-secondary">
-                              Model ID
-                            </label>
-                            <ModelIdInputCompact modelId={mId} onCommit={updateModelId} />
-                          </div>
-                          <div className="flex flex-col gap-1">
-                            <label className="font-body text-[11px] font-medium text-text-secondary">
-                              Model Type
-                            </label>
-                            <select
-                              className={FIELD_CLS}
-                              value={mCfg.type || 'text'}
-                              onChange={(e) => {
-                                const newType = e.target.value as
-                                  | 'text'
-                                  | 'embeddings'
-                                  | 'transcriptions'
-                                  | 'speech'
-                                  | 'image';
-                                if (newType === 'embeddings')
-                                  updateModelConfig(mId, {
-                                    type: newType,
-                                    access_via: ['embeddings'],
-                                  });
-                                else if (newType === 'transcriptions')
-                                  updateModelConfig(mId, {
-                                    type: newType,
-                                    access_via: ['transcriptions'],
-                                  });
-                                else if (newType === 'speech')
-                                  updateModelConfig(mId, { type: newType, access_via: ['speech'] });
-                                else if (newType === 'image')
-                                  updateModelConfig(mId, {
-                                    type: newType,
-                                    access_via: [defaultImageAccess],
-                                  });
-                                else updateModelConfig(mId, { type: newType });
-                              }}
-                            >
-                              <option value="text">Text</option>
-                              <option value="embeddings">Embeddings</option>
-                              <option value="transcriptions">Transcriptions</option>
-                              <option value="speech">Speech</option>
-                              <option value="image">Image</option>
-                            </select>
-                          </div>
-
-                          {(!mCfg.type || mCfg.type === 'text' || mCfg.type === 'image') && (
-                            <div className="flex flex-col gap-1">
-                              <label className="font-body text-[11px] font-medium text-text-secondary">
-                                Access Via
-                              </label>
-                              <div
-                                style={{
-                                  display: 'flex',
-                                  flexWrap: 'wrap',
-                                  gap: '4px',
-                                }}
-                              >
-                                {(mCfg.type === 'image'
-                                  ? imageAccessOptions
-                                  : API_ACCESS_OPTIONS
-                                ).map((option) => {
-                                  const key = apiAccessToKey(option);
-                                  const selected = hasApiAccess(mCfg.access_via, key);
-                                  return (
-                                    <div key={key} className="flex items-center gap-2">
-                                      <label className="flex cursor-pointer items-center gap-[3px]">
-                                        <input
-                                          type="checkbox"
-                                          checked={selected}
-                                          onChange={() => {
-                                            if (mCfg.type === 'image') {
-                                              updateModelConfig(mId, {
-                                                access_via: selected ? [] : [key],
-                                              });
-                                              return;
-                                            }
-
-                                            let next = toggleApiAccess(mCfg.access_via, option);
-                                            if (key === 'responses' && selected) {
-                                              next = next.filter(
-                                                (entry) =>
-                                                  apiAccessToKey(entry) !== 'responses:lite'
-                                              );
-                                            }
-                                            updateModelConfig(mId, { access_via: next });
-                                          }}
-                                        />
-                                        <span
-                                          className="inline-flex items-center rounded-xl font-medium"
-                                          style={{
-                                            ...getApiBadgeStyle(option.type),
-                                            fontSize: '10px',
-                                            padding: '2px 6px',
-                                            opacity: selected ? 1 : 0.5,
-                                          }}
-                                        >
-                                          {option.label}
-                                        </span>
-                                      </label>
-                                      {key === 'responses' && selected && (
-                                        <div className="flex items-center gap-1">
-                                          <label className="flex cursor-pointer items-center gap-1.5 font-body text-[11px] text-text-secondary">
-                                            <input
-                                              type="checkbox"
-                                              checked={hasApiAccess(
-                                                mCfg.access_via,
-                                                'responses:lite'
-                                              )}
-                                              onChange={() => {
-                                                const next = toggleApiAccess(mCfg.access_via, {
-                                                  type: 'responses',
-                                                  subtype: 'lite',
-                                                });
-                                                updateModelConfig(mId, { access_via: next });
-                                              }}
-                                            />
-                                            <span>Lite</span>
-                                          </label>
-                                          <Tooltip
-                                            position="top"
-                                            content={
-                                              <div className="w-64 whitespace-normal font-body leading-relaxed">
-                                                Passes Codex-specific Responses input through
-                                                unchanged, including additional tools. Enable only
-                                                for targets known to support Responses Lite—usually
-                                                direct OpenAI or Codex endpoints. Most
-                                                OpenAI-compatible proxies do not support it.
-                                              </div>
-                                            }
-                                          >
-                                            <button
-                                              type="button"
-                                              aria-label="About Responses Lite"
-                                              className="flex h-4 w-4 items-center justify-center rounded-full text-text-muted transition-colors hover:bg-bg-hover hover:text-text-secondary focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-primary"
-                                            >
-                                              <Info size={12} />
-                                            </button>
-                                          </Tooltip>
-                                        </div>
-                                      )}
-                                    </div>
-                                  );
-                                })}
-                              </div>
-                              {(!mCfg.access_via || mCfg.access_via.length === 0) && (
-                                <span className="font-body text-[11px] text-text-muted italic">
-                                  empty = use any provider API
-                                </span>
-                              )}
-                              {(() => {
-                                const providerBaseUrlMap = getApiBaseUrlMap();
-                                const hasOllamaBaseUrl = Object.entries(providerBaseUrlMap).some(
-                                  ([type, url]) => type === 'ollama' && url && url.trim() !== ''
-                                );
-                                if (hasOllamaBaseUrl && !hasApiAccess(mCfg.access_via, 'ollama')) {
-                                  return (
-                                    <div className="flex items-start gap-2 py-1.5 px-2 bg-info/10 border border-info/30 rounded-sm">
-                                      <Info size={14} className="text-info shrink-0 mt-0.5" />
-                                      <span className="text-[11px] text-info">
-                                        Provider has a native Ollama URL — select{' '}
-                                        <span style={{ fontWeight: 600 }}>ollama</span> above to use
-                                        it.
-                                      </span>
-                                    </div>
-                                  );
-                                }
-                                return null;
-                              })()}
-                              {mCfg.type === 'image' && (
-                                <div className="flex items-start gap-2 py-1.5 px-2 bg-info/10 border border-info/30 rounded-sm">
-                                  <Info size={14} className="text-info shrink-0 mt-0.5" />
-                                  <span className="text-[11px] text-info">
-                                    {isCodexOAuthProvider ? (
-                                      <>
-                                        Codex Images is the only image protocol available on a
-                                        ChatGPT OAuth provider. Requests are signed with the
-                                        provider&apos;s OAuth session, so no base URL is needed.
-                                      </>
-                                    ) : (
-                                      <>
-                                        Choose one image protocol. OpenRouter Images targets the
-                                        dedicated <code>/api/v1/images</code> endpoint;
-                                        OpenAI-compatible and Gemini Images use their native
-                                        adapters.
-                                      </>
-                                    )}
-                                  </span>
-                                </div>
-                              )}
-                              {mCfg.type === 'image' &&
-                                isCodexOAuthProvider &&
-                                (mCfg.access_via?.length ?? 0) > 0 &&
-                                !hasApiAccess(mCfg.access_via, CODEX_IMAGE_ACCESS) && (
-                                  <div className="flex items-start gap-2 py-1.5 px-2 bg-warning/10 border border-warning/30 rounded-sm">
-                                    <AlertTriangle
-                                      size={14}
-                                      className="text-warning shrink-0 mt-0.5"
-                                    />
-                                    <span className="text-[11px] text-warning">
-                                      This model still targets an HTTP image protocol, which a
-                                      ChatGPT OAuth provider cannot serve. Select{' '}
-                                      <span style={{ fontWeight: 600 }}>
-                                        Codex Images (ChatGPT OAuth)
-                                      </span>{' '}
-                                      above.
-                                    </span>
-                                  </div>
-                                )}
-                            </div>
-                          )}
-
-                          {/* pi-ai Model ID */}
-                          <div className="flex flex-col gap-1">
-                            <label className="font-body text-[11px] font-medium text-text-secondary">
-                              pi-ai Model ID
-                            </label>
-                            {piAiProvider && piModels.length > 0 && !piModelCustom[mId] ? (
-                              <select
-                                className={FIELD_CLS}
-                                value={mCfg.pi_ai_model_id ?? ''}
-                                onChange={(e) => {
-                                  const raw = e.target.value;
-                                  if (raw === '__custom__') {
-                                    setPiModelCustom((prev) => ({ ...prev, [mId]: true }));
-                                    return;
-                                  }
-                                  updateModelConfig(mId, {
-                                    pi_ai_model_id: raw || undefined,
-                                  });
-                                }}
-                              >
-                                <option value="">— none —</option>
-                                {piModels.map((m) => (
-                                  <option
-                                    key={m.id}
-                                    value={m.id}
-                                    title={`${m.api}${m.custom ? ' · custom model' : ''}`}
-                                  >
-                                    {m.id}
-                                    {m.custom ? ' (custom)' : ''}
-                                  </option>
-                                ))}
-                                <option value="__custom__">custom...</option>
-                              </select>
-                            ) : (
-                              <div className="flex gap-1">
-                                <input
-                                  className={`${FIELD_CLS} flex-1`}
-                                  type="text"
-                                  placeholder="e.g. gpt-4.1, claude-opus-4-6"
-                                  value={mCfg.pi_ai_model_id ?? ''}
-                                  onChange={(e) => {
-                                    const raw = e.target.value;
-                                    updateModelConfig(mId, {
-                                      pi_ai_model_id: raw || undefined,
-                                    });
-                                  }}
-                                  autoFocus={!!piModelCustom[mId]}
-                                />
-                                {piAiProvider && piModels.length > 0 && piModelCustom[mId] && (
-                                  <button
-                                    type="button"
-                                    className="font-body text-[11px] text-text-muted hover:text-text px-1 flex-shrink-0"
-                                    title="Back to list"
-                                    onClick={() =>
-                                      setPiModelCustom((prev) => ({ ...prev, [mId]: false }))
-                                    }
-                                  >
-                                    ↩
-                                  </button>
-                                )}
-                              </div>
-                            )}
-                          </div>
-                        </div>
-
-                        {/* Right: Pricing Source + Pricing Inputs */}
-                        <div style={{ display: 'flex', flexDirection: 'column', gap: '6px' }}>
-                          <div className="flex flex-col gap-1">
-                            <label className="font-body text-[11px] font-medium text-text-secondary">
-                              Pricing Source
-                            </label>
-                            <select
-                              className={FIELD_CLS}
-                              value={mCfg.pricing?.source || 'simple'}
-                              onChange={(e) => {
-                                const newSource = e.target.value;
-                                let newPricing: any;
-                                if (newSource === 'simple')
-                                  newPricing = {
-                                    source: 'simple',
-                                    input: mCfg.pricing?.input || 0,
-                                    output: mCfg.pricing?.output || 0,
-                                    cached: mCfg.pricing?.cached || 0,
-                                    cache_write: mCfg.pricing?.cache_write || 0,
-                                  };
-                                else if (newSource === 'openrouter')
-                                  newPricing = {
-                                    source: 'openrouter',
-                                    slug: mCfg.pricing?.slug || '',
-                                    ...(mCfg.pricing?.discount !== undefined && {
-                                      discount: mCfg.pricing.discount,
-                                    }),
-                                  };
-                                else if (newSource === 'defined')
-                                  newPricing = {
-                                    source: 'defined',
-                                    range: mCfg.pricing?.range || [],
-                                  };
-                                else if (newSource === 'per_request')
-                                  newPricing = {
-                                    source: 'per_request',
-                                    amount: mCfg.pricing?.amount || 0,
-                                  };
-                                updateModelConfig(mId, { pricing: newPricing });
-                              }}
-                            >
-                              <option value="simple">Simple</option>
-                              <option value="openrouter">OpenRouter</option>
-                              <option value="defined">Ranges (Complex)</option>
-                              <option value="per_request">Per Request (Flat Fee)</option>
-                            </select>
-                          </div>
-
-                          {/* Simple pricing */}
-                          {mCfg.pricing?.source === 'simple' && (
-                            <div
-                              className="grid grid-cols-2"
-                              style={{
-                                background: 'var(--color-bg-subtle)',
-                                padding: '8px',
-                                borderRadius: 'var(--radius-sm)',
-                                gap: '6px',
-                              }}
-                            >
-                              {[
-                                { label: 'Input $/M', key: 'input' },
-                                { label: 'Output $/M', key: 'output' },
-                                { label: 'Cached $/M', key: 'cached' },
-                                { label: 'Cache Write $/M', key: 'cache_write' },
-                              ].map(({ label, key }) => (
-                                <div key={key} className="flex flex-col gap-0.5">
-                                  <label className="font-body text-[11px] font-medium text-text-secondary">
-                                    {label}
-                                  </label>
-                                  <input
-                                    className={FIELD_CLS}
-                                    type="number"
-                                    step="0.000001"
-                                    value={(mCfg.pricing as any)[key] || 0}
-                                    onChange={(e) =>
-                                      updateModelConfig(mId, {
-                                        pricing: {
-                                          ...mCfg.pricing,
-                                          [key]: parseFloat(e.target.value),
-                                        },
-                                      })
-                                    }
-                                  />
-                                </div>
-                              ))}
-                            </div>
-                          )}
-
-                          {/* OpenRouter pricing */}
-                          {mCfg.pricing?.source === 'openrouter' && (
-                            <div
-                              style={{
-                                background: 'var(--color-bg-subtle)',
-                                padding: '8px',
-                                borderRadius: 'var(--radius-sm)',
-                                display: 'flex',
-                                flexDirection: 'column',
-                                gap: '6px',
-                              }}
-                            >
-                              <div className="flex flex-col gap-0.5">
-                                <label className="font-body text-[11px] font-medium text-text-secondary">
-                                  OpenRouter Model Slug
-                                </label>
-                                <OpenRouterSlugInput
-                                  placeholder="e.g. anthropic/claude-3.5-sonnet"
-                                  value={mCfg.pricing.slug || ''}
-                                  onChange={(value) =>
-                                    updateModelConfig(mId, {
-                                      pricing: { ...mCfg.pricing, slug: value },
-                                    })
-                                  }
-                                />
-                              </div>
-                              <div className="flex flex-col gap-0.5">
-                                <label className="font-body text-[11px] font-medium text-text-secondary">
-                                  Discount{' '}
-                                  <span className="font-normal text-text-muted">
-                                    (0.1 = 10% off)
-                                  </span>
-                                </label>
-                                <input
-                                  className="w-full py-1 px-2 font-body text-[12px] leading-none text-text bg-bg-glass border border-border-glass rounded-sm outline-none focus:border-primary"
-                                  type="number"
-                                  step="0.01"
-                                  min="0"
-                                  max="1"
-                                  value={mCfg.pricing.discount ?? ''}
-                                  onChange={(e) => {
-                                    const val = e.target.value;
-                                    if (val === '') {
-                                      const { discount, ...rest } = mCfg.pricing;
-                                      updateModelConfig(mId, { pricing: rest });
-                                    } else
-                                      updateModelConfig(mId, {
-                                        pricing: { ...mCfg.pricing, discount: parseFloat(val) },
-                                      });
-                                  }}
-                                />
-                              </div>
-                            </div>
-                          )}
-
-                          {/* Defined/ranges pricing */}
-                          {mCfg.pricing?.source === 'defined' && (
-                            <div
-                              style={{
-                                background: 'var(--color-bg-subtle)',
-                                padding: '8px',
-                                borderRadius: 'var(--radius-sm)',
-                                display: 'flex',
-                                flexDirection: 'column',
-                                gap: '8px',
-                              }}
-                            >
-                              <div
-                                style={{
-                                  display: 'flex',
-                                  justifyContent: 'space-between',
-                                  alignItems: 'center',
-                                }}
-                              >
-                                <span className="font-body text-[11px] font-medium text-text-secondary">
-                                  Pricing Ranges
-                                </span>
-                                <Button
-                                  size="sm"
-                                  variant="secondary"
-                                  onClick={() => {
-                                    const currentRanges = mCfg.pricing.range || [];
-                                    updateModelConfig(mId, {
-                                      pricing: {
-                                        ...mCfg.pricing,
-                                        range: [
-                                          ...currentRanges,
-                                          {
-                                            lower_bound: 0,
-                                            upper_bound: 0,
-                                            input_per_m: 0,
-                                            output_per_m: 0,
-                                            cache_write_per_m: 0,
-                                          },
-                                        ],
-                                      },
-                                    });
-                                  }}
-                                  leftIcon={<Plus size={14} />}
-                                >
-                                  Add Range
-                                </Button>
-                              </div>
-                              {(mCfg.pricing.range || []).map((range: any, idx: number) => (
-                                <div
-                                  key={idx}
-                                  style={{
-                                    border: '1px solid var(--color-border-glass)',
-                                    padding: '8px',
-                                    borderRadius: 'var(--radius-sm)',
-                                    position: 'relative',
-                                    display: 'flex',
-                                    flexDirection: 'column',
-                                    gap: '6px',
-                                  }}
-                                >
-                                  <Button
-                                    size="sm"
-                                    variant="ghost"
-                                    style={{
-                                      position: 'absolute',
-                                      top: '6px',
-                                      right: '6px',
-                                      color: 'var(--color-danger)',
-                                      padding: '4px',
-                                    }}
-                                    onClick={() => {
-                                      const r = [...mCfg.pricing.range];
-                                      r.splice(idx, 1);
-                                      updateModelConfig(mId, {
-                                        pricing: { ...mCfg.pricing, range: r },
-                                      });
-                                    }}
-                                  >
-                                    <X size={14} />
-                                  </Button>
-                                  <div className="grid grid-cols-2" style={{ gap: '6px' }}>
-                                    {[
-                                      {
-                                        label: 'Lower Bound',
-                                        field: 'lower_bound',
-                                        val: range.lower_bound,
-                                      },
-                                      {
-                                        label: 'Upper Bound (0=∞)',
-                                        field: 'upper_bound',
-                                        val: range.upper_bound === Infinity ? 0 : range.upper_bound,
-                                      },
-                                      {
-                                        label: 'Input $/M',
-                                        field: 'input_per_m',
-                                        val: range.input_per_m,
-                                      },
-                                      {
-                                        label: 'Output $/M',
-                                        field: 'output_per_m',
-                                        val: range.output_per_m,
-                                      },
-                                      {
-                                        label: 'Cached $/M',
-                                        field: 'cached_per_m',
-                                        val: range.cached_per_m || 0,
-                                      },
-                                      {
-                                        label: 'Cache Write $/M',
-                                        field: 'cache_write_per_m',
-                                        val: range.cache_write_per_m || 0,
-                                      },
-                                    ].map(({ label, field, val }) => (
-                                      <div key={field} className="flex flex-col gap-0.5">
-                                        <label className="font-body text-[10px] font-medium text-text-secondary">
-                                          {label}
-                                        </label>
-                                        <input
-                                          className="w-full py-1 px-2 font-body text-[12px] leading-none text-text bg-bg-glass border border-border-glass rounded-sm outline-none focus:border-primary"
-                                          type="number"
-                                          step="0.000001"
-                                          value={val}
-                                          onChange={(e) => {
-                                            const r = [...mCfg.pricing.range];
-                                            const v =
-                                              field === 'upper_bound'
-                                                ? parseFloat(e.target.value) === 0
-                                                  ? Infinity
-                                                  : parseFloat(e.target.value)
-                                                : parseFloat(e.target.value);
-                                            r[idx] = {
-                                              ...range,
-                                              [field]: Number.isFinite(v)
-                                                ? v
-                                                : field === 'upper_bound'
-                                                  ? Infinity
-                                                  : 0,
-                                            };
-                                            updateModelConfig(mId, {
-                                              pricing: { ...mCfg.pricing, range: r },
-                                            });
-                                          }}
-                                        />
-                                      </div>
-                                    ))}
-                                  </div>
-                                </div>
-                              ))}
-                              {(!mCfg.pricing.range || mCfg.pricing.range.length === 0) && (
-                                <div className="text-text-muted italic text-center text-[11px] py-2">
-                                  No ranges defined.
-                                </div>
-                              )}
-                            </div>
-                          )}
-
-                          {/* Per Request pricing */}
-                          {mCfg.pricing?.source === 'per_request' && (
-                            <div
-                              style={{
-                                background: 'var(--color-bg-subtle)',
-                                padding: '8px',
-                                borderRadius: 'var(--radius-sm)',
-                                display: 'flex',
-                                flexDirection: 'column',
-                                gap: '6px',
-                              }}
-                            >
-                              <div className="flex flex-col gap-0.5">
-                                <label className="font-body text-[11px] font-medium text-text-secondary">
-                                  Cost Per Request ($)
-                                </label>
-                                <input
-                                  className="w-full py-1 px-2 font-body text-[12px] leading-none text-text bg-bg-glass border border-border-glass rounded-sm outline-none focus:border-primary"
-                                  type="number"
-                                  step="0.000001"
-                                  min="0"
-                                  value={mCfg.pricing.amount || 0}
-                                  onChange={(e) =>
-                                    updateModelConfig(mId, {
-                                      pricing: {
-                                        ...mCfg.pricing,
-                                        amount: parseFloat(e.target.value) || 0,
-                                      },
-                                    })
-                                  }
-                                />
-                              </div>
-                              <span className="font-body text-[11px] text-text-muted italic">
-                                Flat fee per API call, regardless of token count.
-                              </span>
-                            </div>
-                          )}
-                        </div>
-                      </div>
-
-                      {/* Per-Model Adapters — disclosure */}
-                      <div className="border border-border-glass rounded-md overflow-hidden">
-                        <div
-                          className="p-2 px-3 flex items-center gap-2 cursor-pointer bg-bg-hover hover:bg-bg-glass"
-                          onClick={() =>
-                            setModelAdaptersOpen((prev) => ({ ...prev, [mId]: !prev[mId] }))
-                          }
-                        >
-                          {modelAdaptersOpen[mId] ? (
-                            <ChevronDown size={14} />
-                          ) : (
-                            <ChevronRight size={14} />
-                          )}
-                          <span className="font-body text-[12px] font-medium text-text-secondary flex-1">
-                            Model Adapters
-                          </span>
-                          {(() => {
-                            const modelAdapters: any[] = mCfg.adapter
-                              ? Array.isArray(mCfg.adapter)
-                                ? mCfg.adapter
-                                : [mCfg.adapter]
-                              : [];
-                            return modelAdapters.length > 0 ? (
-                              <Badge
-                                status="neutral"
-                                style={{ fontSize: '10px', padding: '2px 8px' }}
-                              >
-                                {modelAdapters.length}
-                              </Badge>
-                            ) : null;
-                          })()}
-                        </div>
-                        {modelAdaptersOpen[mId] && (
-                          <div
-                            style={{
-                              display: 'grid',
-                              gridTemplateColumns: '1fr 1fr',
-                              gap: '4px',
-                              padding: '8px',
-                              borderTop: '1px solid var(--color-border-glass)',
-                              background: 'var(--color-bg-subtle)',
-                            }}
-                          >
-                            {isGpt5Model(mId) &&
-                              (() => {
-                                const modelAdapters: any[] = mCfg.adapter
-                                  ? Array.isArray(mCfg.adapter)
-                                    ? mCfg.adapter
-                                    : [mCfg.adapter]
-                                  : [];
-                                const suppressionDisabled = modelAdapters.some(
-                                  (entry: any) =>
-                                    typeof entry !== 'string' &&
-                                    entry.name === GPT5_SUPPRESSION_ADAPTER &&
-                                    entry.enabled === false
-                                );
-                                return (
-                                  <label
-                                    style={{
-                                      gridColumn: '1 / -1',
-                                      display: 'flex',
-                                      alignItems: 'flex-start',
-                                      gap: '8px',
-                                      cursor: 'pointer',
-                                      padding: '4px 8px',
-                                      borderRadius: 'var(--radius-sm)',
-                                      border: '1px solid var(--color-border-glass)',
-                                      background: suppressionDisabled
-                                        ? 'var(--color-bg-glass)'
-                                        : 'var(--color-bg-hover)',
-                                    }}
-                                  >
-                                    <input
-                                      type="checkbox"
-                                      checked={!suppressionDisabled}
-                                      style={{ marginTop: '2px', flexShrink: 0 }}
-                                      onChange={() => {
-                                        const withoutSuppression = modelAdapters.filter(
-                                          (entry: any) =>
-                                            (typeof entry === 'string' ? entry : entry.name) !==
-                                            GPT5_SUPPRESSION_ADAPTER
-                                        );
-                                        const next = suppressionDisabled
-                                          ? withoutSuppression
-                                          : [
-                                              ...withoutSuppression,
-                                              {
-                                                name: GPT5_SUPPRESSION_ADAPTER,
-                                                options: {},
-                                                enabled: false,
-                                              },
-                                            ];
-                                        updateModelConfig(mId, {
-                                          adapter: next.length > 0 ? next : undefined,
-                                        });
-                                      }}
-                                    />
-                                    <div>
-                                      <div className="font-body text-[12px] font-medium text-text">
-                                        Suppress Unsupported GPT-5 Options
-                                      </div>
-                                      <div
-                                        className="font-body text-[11px] text-text-secondary"
-                                        style={{ lineHeight: 1.35 }}
-                                      >
-                                        Enabled by default. Removes generation options GPT-5 does
-                                        not accept.
-                                      </div>
-                                    </div>
-                                  </label>
-                                );
-                              })()}
-                            {KNOWN_ADAPTERS.map((a) => {
-                              const modelAdapters: any[] = mCfg.adapter
-                                ? Array.isArray(mCfg.adapter)
-                                  ? mCfg.adapter
-                                  : [mCfg.adapter]
-                                : [];
-                              const active = modelAdapters.some(
-                                (e: any) => (typeof e === 'string' ? e : e.name) === a.value
-                              );
-                              return (
-                                <label
-                                  key={a.value}
-                                  style={{
-                                    display: 'flex',
-                                    alignItems: 'flex-start',
-                                    gap: '8px',
-                                    cursor: 'pointer',
-                                    padding: '4px 8px',
-                                    borderRadius: 'var(--radius-sm)',
-                                    border: '1px solid var(--color-border-glass)',
-                                    background: active
-                                      ? 'var(--color-bg-hover)'
-                                      : 'var(--color-bg-glass)',
-                                  }}
-                                >
-                                  <input
-                                    type="checkbox"
-                                    checked={active}
-                                    style={{ marginTop: '2px', flexShrink: 0 }}
-                                    onChange={() => {
-                                      const modelAdapters: any[] = mCfg.adapter
-                                        ? Array.isArray(mCfg.adapter)
-                                          ? mCfg.adapter
-                                          : [mCfg.adapter]
-                                        : [];
-                                      const next = active
-                                        ? modelAdapters.filter(
-                                            (e: any) =>
-                                              (typeof e === 'string' ? e : e.name) !== a.value
-                                          )
-                                        : [
-                                            ...modelAdapters,
-                                            {
-                                              name: a.value,
-                                              options:
-                                                a.value === 'model_override'
-                                                  ? { rules: [] }
-                                                  : a.value === 'reasoning_rewrite'
-                                                    ? { rules: [] }
-                                                    : {},
-                                            },
-                                          ];
-                                      updateModelConfig(mId, {
-                                        adapter: next.length > 0 ? next : undefined,
-                                      });
-                                    }}
-                                  />
-                                  <div>
-                                    <div className="font-body text-[12px] font-medium text-text">
-                                      {a.label}
-                                    </div>
-                                    <div
-                                      className="font-body text-[11px] text-text-secondary"
-                                      style={{ lineHeight: 1.35 }}
-                                    >
-                                      {a.description}
-                                    </div>
-                                  </div>
-                                </label>
-                              );
-                            })}
-                            {/* model_override rules editor */}
-                            {(() => {
-                              const modelAdapters: any[] = mCfg.adapter
-                                ? Array.isArray(mCfg.adapter)
-                                  ? mCfg.adapter
-                                  : [mCfg.adapter]
-                                : [];
-                              const overrideEntry = modelAdapters.find(
-                                (e: any) =>
-                                  (typeof e === 'string' ? e : e.name) === 'model_override'
-                              );
-                              if (!overrideEntry || typeof overrideEntry === 'string') return null;
-                              const rules: any[] = overrideEntry.options?.rules ?? [];
-                              return (
-                                <div
-                                  style={{
-                                    gridColumn: '1 / -1',
-                                    borderTop: '1px solid var(--color-border-glass)',
-                                    marginTop: '4px',
-                                    paddingTop: '6px',
-                                  }}
-                                >
-                                  <div className="font-body text-[11px] font-medium text-text-secondary mb-1">
-                                    Model Override Rules
-                                  </div>
-                                  <div
-                                    className="font-body text-[10px] text-text-muted mb-2"
-                                    style={{ lineHeight: 1.3 }}
-                                  >
-                                    When ANY condition matches, rewrite the model name. Use dotted
-                                    paths like reasoning.enabled.
-                                  </div>
-                                  {rules.map((rule: any, rIdx: number) => (
-                                    <div
-                                      key={rIdx}
-                                      style={{
-                                        border: '1px solid var(--color-border-glass)',
-                                        borderRadius: 'var(--radius-sm)',
-                                        padding: '6px',
-                                        marginBottom: '4px',
-                                        background: 'var(--color-bg-subtle)',
-                                      }}
-                                    >
-                                      {/* Rewrite rule */}
-                                      <div className="font-body text-[10px] font-medium text-text-muted mb-1">
-                                        Rewrite
-                                      </div>
-                                      <div
-                                        style={{
-                                          display: 'flex',
-                                          gap: '4px',
-                                          alignItems: 'center',
-                                        }}
-                                      >
-                                        <div
-                                          className="font-body text-[12px] text-text-muted"
-                                          style={{
-                                            flex: 2,
-                                            padding: '5px 8px',
-                                            background: 'var(--color-bg-glass)',
-                                            border: '1px solid var(--color-border-glass)',
-                                            borderRadius: 'var(--radius-sm)',
-                                            overflow: 'hidden',
-                                            textOverflow: 'ellipsis',
-                                            whiteSpace: 'nowrap',
-                                          }}
-                                          title={mId}
-                                        >
-                                          {mId}
-                                        </div>
-                                        <span className="font-body text-[11px] text-text-muted">
-                                          →
-                                        </span>
-                                        <div style={{ flex: 1 }}>
-                                          <DebouncedInput
-                                            placeholder="Rewrite to (e.g. deepseek-r1-fast)"
-                                            value={rule.rewriteTo ?? ''}
-                                            onChange={(val: string) => {
-                                              const updated = [...rules];
-                                              updated[rIdx] = {
-                                                ...updated[rIdx],
-                                                rewriteTo: val,
-                                              };
-                                              const newAdapters = modelAdapters.map((entry: any) =>
-                                                typeof entry !== 'string' &&
-                                                entry.name === 'model_override'
-                                                  ? {
-                                                      ...entry,
-                                                      options: { ...entry.options, rules: updated },
-                                                    }
-                                                  : entry
-                                              );
-                                              updateModelConfig(mId, { adapter: newAdapters });
-                                            }}
-                                          />
-                                        </div>
-                                        <Button
-                                          variant="ghost"
-                                          size="sm"
-                                          onClick={() => {
-                                            const updated = rules.filter(
-                                              (_: any, i: number) => i !== rIdx
-                                            );
-                                            const newAdapters = modelAdapters.map((entry: any) =>
-                                              typeof entry !== 'string' &&
-                                              entry.name === 'model_override'
-                                                ? {
-                                                    ...entry,
-                                                    options: { ...entry.options, rules: updated },
-                                                  }
-                                                : entry
-                                            );
-                                            updateModelConfig(mId, { adapter: newAdapters });
-                                          }}
-                                          style={{ padding: '4px' }}
-                                        >
-                                          <Trash2
-                                            size={14}
-                                            style={{ color: 'var(--color-danger)' }}
-                                          />
-                                        </Button>
-                                      </div>
-                                      {/* Conditions separator */}
-                                      <div
-                                        style={{
-                                          borderTop: '1px solid var(--color-border-glass)',
-                                          margin: '6px 0 4px 0',
-                                        }}
-                                      />
-                                      <div className="font-body text-[10px] font-medium text-text-muted mb-1">
-                                        Conditions (any match triggers rewrite)
-                                      </div>
-                                      {/* Condition column headers */}
-                                      <div
-                                        style={{
-                                          display: 'flex',
-                                          gap: '4px',
-                                          marginBottom: '2px',
-                                          marginLeft: '8px',
-                                        }}
-                                      >
-                                        <span
-                                          className="font-body text-[9px] font-medium text-text-muted"
-                                          style={{ flex: 1, paddingLeft: '8px' }}
-                                        >
-                                          Field path (dotted)
-                                        </span>
-                                        <span
-                                          className="font-body text-[9px] font-medium text-text-muted"
-                                          style={{ flex: 1, paddingLeft: '8px' }}
-                                        >
-                                          Value (blank = presence check)
-                                        </span>
-                                        {/* spacer for delete button column */}
-                                        <span style={{ width: '28px' }} />
-                                      </div>
-                                      {/* Conditions */}
-                                      {(rule.conditions ?? []).map((cond: any, cIdx: number) => (
-                                        <div
-                                          key={cIdx}
-                                          style={{
-                                            display: 'flex',
-                                            gap: '4px',
-                                            marginBottom: '2px',
-                                            marginLeft: '8px',
-                                          }}
-                                        >
-                                          <div style={{ flex: 2 }}>
-                                            <DebouncedInput
-                                              placeholder="e.g. reasoning.enabled"
-                                              value={cond.field ?? ''}
-                                              onChange={(val: string) => {
-                                                const updated = [...rules];
-                                                const newConditions = [...updated[rIdx].conditions];
-                                                newConditions[cIdx] = {
-                                                  ...newConditions[cIdx],
-                                                  field: val,
-                                                };
-                                                updated[rIdx] = {
-                                                  ...updated[rIdx],
-                                                  conditions: newConditions,
-                                                };
-                                                const newAdapters = modelAdapters.map(
-                                                  (entry: any) =>
-                                                    typeof entry !== 'string' &&
-                                                    entry.name === 'model_override'
-                                                      ? {
-                                                          ...entry,
-                                                          options: {
-                                                            ...entry.options,
-                                                            rules: updated,
-                                                          },
-                                                        }
-                                                      : entry
-                                                );
-                                                updateModelConfig(mId, { adapter: newAdapters });
-                                              }}
-                                            />
-                                          </div>
-                                          <div style={{ flex: 1 }}>
-                                            <DebouncedInput
-                                              placeholder="e.g. false, 0, none"
-                                              value={
-                                                cond.value !== undefined ? String(cond.value) : ''
-                                              }
-                                              onChange={(val: string) => {
-                                                const parsed =
-                                                  val === ''
-                                                    ? undefined
-                                                    : val === 'true'
-                                                      ? true
-                                                      : val === 'false'
-                                                        ? false
-                                                        : isNaN(Number(val))
-                                                          ? val
-                                                          : Number(val);
-                                                const updated = [...rules];
-                                                const newConditions = [...updated[rIdx].conditions];
-                                                newConditions[cIdx] = {
-                                                  field: newConditions[cIdx].field,
-                                                  value: parsed,
-                                                };
-                                                updated[rIdx] = {
-                                                  ...updated[rIdx],
-                                                  conditions: newConditions,
-                                                };
-                                                const newAdapters = modelAdapters.map(
-                                                  (entry: any) =>
-                                                    typeof entry !== 'string' &&
-                                                    entry.name === 'model_override'
-                                                      ? {
-                                                          ...entry,
-                                                          options: {
-                                                            ...entry.options,
-                                                            rules: updated,
-                                                          },
-                                                        }
-                                                      : entry
-                                                );
-                                                updateModelConfig(mId, { adapter: newAdapters });
-                                              }}
-                                            />
-                                          </div>
-                                          <Button
-                                            variant="ghost"
-                                            size="sm"
-                                            onClick={() => {
-                                              const updated = [...rules];
-                                              const newConditions = updated[rIdx].conditions.filter(
-                                                (_: any, i: number) => i !== cIdx
-                                              );
-                                              updated[rIdx] = {
-                                                ...updated[rIdx],
-                                                conditions: newConditions,
-                                              };
-                                              const newAdapters = modelAdapters.map((entry: any) =>
-                                                typeof entry !== 'string' &&
-                                                entry.name === 'model_override'
-                                                  ? {
-                                                      ...entry,
-                                                      options: { ...entry.options, rules: updated },
-                                                    }
-                                                  : entry
-                                              );
-                                              updateModelConfig(mId, { adapter: newAdapters });
-                                            }}
-                                            style={{ padding: '4px' }}
-                                          >
-                                            <Trash2
-                                              size={12}
-                                              style={{ color: 'var(--color-danger)' }}
-                                            />
-                                          </Button>
-                                        </div>
-                                      ))}
-                                      <Button
-                                        variant="ghost"
-                                        size="sm"
-                                        onClick={() => {
-                                          const updated = [...rules];
-                                          updated[rIdx] = {
-                                            ...updated[rIdx],
-                                            conditions: [
-                                              ...(updated[rIdx].conditions ?? []),
-                                              { field: '' },
-                                            ],
-                                          };
-                                          const newAdapters = modelAdapters.map((entry: any) =>
-                                            typeof entry !== 'string' &&
-                                            entry.name === 'model_override'
-                                              ? {
-                                                  ...entry,
-                                                  options: { ...entry.options, rules: updated },
-                                                }
-                                              : entry
-                                          );
-                                          updateModelConfig(mId, { adapter: newAdapters });
-                                        }}
-                                        style={{ marginLeft: '8px', padding: '2px 6px' }}
-                                      >
-                                        <Plus size={12} />{' '}
-                                        <span className="font-body text-[10px]">Condition</span>
-                                      </Button>
-                                    </div>
-                                  ))}
-                                  <Button
-                                    variant="secondary"
-                                    size="sm"
-                                    onClick={() => {
-                                      const newRule = {
-                                        model: mId,
-                                        rewriteTo: '',
-                                        conditions: [{ field: '' }],
-                                      };
-                                      const updated = [...rules, newRule];
-                                      const newAdapters = modelAdapters.map((entry: any) =>
-                                        typeof entry !== 'string' && entry.name === 'model_override'
-                                          ? {
-                                              ...entry,
-                                              options: { ...entry.options, rules: updated },
-                                            }
-                                          : entry
-                                      );
-                                      updateModelConfig(mId, { adapter: newAdapters });
-                                    }}
-                                    style={{ marginTop: '2px' }}
-                                  >
-                                    <Plus size={12} />{' '}
-                                    <span className="font-body text-[10px]">Rule</span>
-                                  </Button>
-                                </div>
-                              );
-                            })()}
-                            {/* reasoning_rewrite rules editor (shared component) */}
-                            {(() => {
-                              const modelAdapters: any[] = mCfg.adapter
-                                ? Array.isArray(mCfg.adapter)
-                                  ? mCfg.adapter
-                                  : [mCfg.adapter]
-                                : [];
-                              return (
-                                <ReasoningRewriteRulesEditor
-                                  adapters={modelAdapters}
-                                  onChange={(next: any[]) =>
-                                    updateModelConfig(mId, { adapter: next })
-                                  }
-                                />
-                              );
-                            })()}
-                          </div>
-                        )}
-                      </div>
-
-                      {/* Per-Model Extra Body Fields */}
-                      <div className="border border-border-glass rounded-md overflow-hidden">
-                        <div
-                          className="p-2 px-3 flex items-center gap-2 cursor-pointer bg-bg-hover hover:bg-bg-glass"
-                          onClick={() =>
-                            setIsModelExtraBodyOpen({
-                              ...isModelExtraBodyOpen,
-                              [mId]: !isModelExtraBodyOpen[mId],
-                            })
-                          }
-                        >
-                          {isModelExtraBodyOpen[mId] ? (
-                            <ChevronDown size={14} />
-                          ) : (
-                            <ChevronRight size={14} />
-                          )}
-                          <span className="font-body text-[12px] font-medium text-text-secondary flex-1">
-                            Extra Body Fields
-                          </span>
-                          <Badge status="neutral" style={{ fontSize: '10px', padding: '2px 8px' }}>
-                            {Object.keys(mCfg.extraBody || {}).length}
-                          </Badge>
-                          <Button
-                            size="sm"
-                            variant="secondary"
-                            style={{ padding: '2px 6px', lineHeight: 1 }}
-                            onClick={(e) => {
-                              e.stopPropagation();
-                              addModelKV(mId);
-                              setIsModelExtraBodyOpen({ ...isModelExtraBodyOpen, [mId]: true });
-                            }}
-                          >
-                            <Plus size={14} />
-                          </Button>
-                        </div>
-                        {isModelExtraBodyOpen[mId] && (
-                          <div
-                            style={{
-                              display: 'flex',
-                              flexDirection: 'column',
-                              gap: '4px',
-                              padding: '8px',
-                              borderTop: '1px solid var(--color-border-glass)',
-                              background: 'var(--color-bg-subtle)',
-                            }}
-                          >
-                            {Object.entries(mCfg.extraBody || {}).length === 0 && (
-                              <div className="font-body text-[11px] text-text-secondary italic">
-                                No extra body fields configured.
-                              </div>
-                            )}
-                            {Object.entries(mCfg.extraBody || {}).map(
-                              ([key, val]: [string, any], idx: number) => (
-                                <div key={idx} style={{ display: 'flex', gap: '6px' }}>
-                                  <DebouncedInput
-                                    placeholder="Field Name"
-                                    value={key}
-                                    onChange={(newKey: string) =>
-                                      updateModelKV(mId, key, newKey, val)
-                                    }
-                                    style={{ flex: 1 }}
-                                  />
-                                  <DebouncedInput
-                                    placeholder="Value"
-                                    value={
-                                      typeof val === 'object' ? JSON.stringify(val) : String(val)
-                                    }
-                                    onChange={(val: string) => {
-                                      try {
-                                        updateModelKV(mId, key, key, JSON.parse(val));
-                                      } catch {
-                                        updateModelKV(mId, key, key, val);
-                                      }
-                                    }}
-                                    style={{ flex: 1 }}
-                                  />
-                                  <Button
-                                    variant="ghost"
-                                    size="sm"
-                                    onClick={() => removeModelKV(mId, key)}
-                                    style={{ padding: '4px' }}
-                                  >
-                                    <Trash2 size={14} style={{ color: 'var(--color-danger)' }} />
-                                  </Button>
-                                </div>
-                              )
-                            )}
-                          </div>
-                        )}
-                      </div>
-
-                      {/* Per-Model Advanced */}
-                      <div className="border border-border-glass rounded-md overflow-hidden">
-                        <div
-                          className="p-2 px-3 flex items-center gap-2 cursor-pointer bg-bg-hover hover:bg-bg-glass"
-                          onClick={() =>
-                            setModelAdvancedOpen((prev) => ({ ...prev, [mId]: !prev[mId] }))
-                          }
-                        >
-                          {modelAdvancedOpen[mId] ? (
-                            <ChevronDown size={14} />
-                          ) : (
-                            <ChevronRight size={14} />
-                          )}
-                          <span className="font-body text-[12px] font-medium text-text-secondary flex-1">
-                            Advanced
-                          </span>
-                          {mCfg.maxConcurrency != null && (
-                            <Badge
-                              status="neutral"
-                              style={{ fontSize: '10px', padding: '2px 8px' }}
-                            >
-                              Concurrency: {mCfg.maxConcurrency}
-                            </Badge>
-                          )}
-                          {mCfg.auto_compat === true && (
-                            <Badge
-                              status="neutral"
-                              style={{ fontSize: '10px', padding: '2px 8px' }}
-                            >
-                              Auto Compat
-                            </Badge>
-                          )}
-                        </div>
-                        {modelAdvancedOpen[mId] && (
-                          <div
-                            style={{
-                              display: 'flex',
-                              flexDirection: 'column',
-                              gap: '6px',
-                              padding: '8px',
-                              borderTop: '1px solid var(--color-border-glass)',
-                              background: 'var(--color-bg-subtle)',
-                            }}
-                          >
-                            <div className="flex flex-col gap-0.5">
-                              <label className="flex items-start gap-2 py-1 cursor-pointer">
-                                <input
-                                  type="checkbox"
-                                  checked={mCfg.auto_compat === true}
-                                  onChange={(e) =>
-                                    updateModelConfig(mId, {
-                                      auto_compat: e.target.checked ? true : undefined,
-                                    })
-                                  }
-                                />
-                                <div>
-                                  <div className="font-body text-[12px] text-text">Auto Compat</div>
-                                  <div
-                                    className="font-body text-[11px] text-text-muted"
-                                    style={{ lineHeight: 1.35 }}
-                                  >
-                                    Use pi-ai registry hints for this model.
-                                  </div>
-                                </div>
-                              </label>
-                            </div>
-                            <div className="flex flex-col gap-0.5">
-                              <label className="font-body text-[11px] font-medium text-text-secondary">
-                                Max Concurrency
-                                <span className="font-normal text-[10px] text-text-muted ml-1">
-                                  this model only
-                                </span>
-                              </label>
-                              <input
-                                className={FIELD_CLS}
-                                type="number"
-                                step="1"
-                                min="1"
-                                placeholder="No limit"
-                                value={mCfg.maxConcurrency != null ? mCfg.maxConcurrency : ''}
-                                onChange={(e) => {
-                                  const raw = e.target.value;
-                                  if (raw === '') {
-                                    updateModelConfig(mId, { maxConcurrency: undefined });
-                                  } else {
-                                    const val = Number(raw);
-                                    if (Number.isFinite(val) && val >= 1) {
-                                      updateModelConfig(mId, { maxConcurrency: val });
-                                    }
-                                  }
-                                }}
-                              />
-                              <span className="font-body text-[11px] text-text-muted italic">
-                                Limit in-flight requests for this model. Leave empty to use the
-                                provider-wide limit or no limit.
-                              </span>
-                            </div>
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  )}
-                </div>
-              );
-            })}
-          </div>
-        </div>
-      )}
-    </div>
+    <ModelList
+      editingProvider={editingProvider}
+      isModelsOpen={isModelsOpen}
+      setIsModelsOpen={setIsModelsOpen}
+      openModelIdx={openModelIdx}
+      setOpenModelIdx={setOpenModelIdx}
+      isModelExtraBodyOpen={isModelExtraBodyOpen}
+      setIsModelExtraBodyOpen={setIsModelExtraBodyOpen}
+      testStates={testStates}
+      addModel={addModel}
+      updateModelConfig={updateModelConfig}
+      addModelKV={addModelKV}
+      updateModelKV={updateModelKV}
+      removeModelKV={removeModelKV}
+      onOpenFetchModels={onOpenFetchModels}
+      onTestModel={onTestModel}
+      onDismissTestMessage={onDismissTestMessage}
+      getApiBaseUrlMap={getApiBaseUrlMap}
+      isNewProvider={isNewProvider}
+      modelAdaptersOpen={modelAdaptersOpen}
+      setModelAdaptersOpen={setModelAdaptersOpen}
+      modelAdvancedOpen={modelAdvancedOpen}
+      setModelAdvancedOpen={setModelAdvancedOpen}
+      piAiProvider={editingProvider.pi_ai_provider}
+      piModels={piAiModels.piModels}
+      piModelCustom={piAiModels.piModelCustom}
+      setPiModelCustom={piAiModels.setPiModelCustom}
+      onUpdateModelId={handleUpdateModelId}
+      onRemoveModel={handleRemoveModel}
+      isCodexOAuthProvider={isCodexOAuthProvider}
+    />
   );
 }
