+++ packages/frontend/src/components/providers/model-editor/types.ts
@@ -1,0 +1,9 @@
+export type ModelConfig = Record<string, any>;
+
+export interface ModelTestState {
+  loading: boolean;
+  result?: 'success' | 'error';
+  message?: string;
+  showResult: boolean;
+  showMessage?: boolean;
+}
